allow for more flags in loop lines in case they're needed

+++ lib/src/cmd_private.h
@@ -82,7 +82,7 @@ struct stmt_ {
     gint16 ci;     /* command index */
     gint16 next;   /* line index to skip to after loop or conditional */
     gint16 idx;    /* 1-based line index allowing for blanks (functions only) */
-    guint8 flags;  /* state flags */
+    gint16 flags;  /* state flags */
     void *ptr;     /* attachment for LOOPSET or GENERATOR */
 };
